split out container, fix warnings
